pyproject.toml
[project]
name = "qinglongmy"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "jieba3k>=0.35.1",
    "markdown>=3.10",
    "pysnowball>=0.1.8",
    "python-dotenv>=1.2.1",
    "pytrends>=4.9.2",
    "requests>=2.32.5",
]
